+++ R/data_plot.R
@@ -22,6 +22,7 @@
 #'   (using \code{\link[bayestestR]{simulate_prior}}) and added to the plot.
 #' @param priors_alpha Alpha value of the prior distributions.
 #' @param point_size Size of point-geoms.
+#' @param digits Number of decimals used for values.
 #' @param rope_alpha,si_alpha Transparency level of ROPE/SI ribbon.
 #' @param rope_color,si_color Color of ROPE/SI ribbon.
 #' @param support_only Plot only the support data, or show the "raw" prior and posterior distributions? Only applies when plotting \code{\link[bayestestR]{si}}.
@@ -42,6 +43,9 @@
 #' @param log Show log-transformed Bayes factors.
 #' @param text_size Size of text labels.
 #' @param text_color Color of text labels.
+#' @param show_values Logical, if \code{TRUE}, values are displayed.
+#' @param show_p Logical, if \code{TRUE}, p-values or significant level is displayed.
+#' @param show_legend Logical, show or hide legend.
 #' @param threshold_coefficient Numeric, threshold at which value coefficients will be displayed.
 #' @param threshold_p Numeric, threshold at which value p-values will be displayed.
 #' @param ci Logical, whether confidence intervals should be added to the plot.
@@ -68,44 +72,8 @@
 #' \cr \cr
 #' Most \code{plot()}-functions work out-of-the-box, i.e. you don't need to do
 #' much more than calling \code{plot(<object>)} (see 'Examples'). Some plot-functions
-#' allow to specify arguments to modify the transparancy or color of geoms, these
+#' allow to specify arguments to modify the transparency or color of geoms, these
 #' are shown in the 'Usage' section.
-#' \cr \cr
-#' Plot-functions are available for objects from following functions (note that
-#' functions from packages might be listed here that are currently still in
-#' development and probably not yet available):
-#' \itemize{
-#'   \item \code{bayestestR::bayesfactor_models()}
-#'   \item \code{bayestestR::bayesfactor_parameters()}
-#'   \item \code{bayestestR::ci()}
-#'   \item \code{bayestestR::equivalence_test()}
-#'   \item \code{bayestestR::estimate_density()}
-#'   \item \code{bayestestR::eti()}
-#'   \item \code{bayestestR::hdi()}
-#'   \item \code{bayestestR::map_estimate()}
-#'   \item \code{bayestestR::p_direction()}
-#'   \item \code{bayestestR::p_significance()}
-#'   \item \code{bayestestR::point_estimate()}
-#'   \item \code{bayestestR::rope()}
-#'   \item \code{bayestestR::si()}
-#'   \item \code{correlation::correlation()}
-#'   \item \code{modelbased::estimate_contrasts()}
-#'   \item \code{parameters::cluster_analysis()}
-#'   \item \code{parameters::model_parameters()}
-#'   \item \code{parameters::n_factors()}
-#'   \item \code{parameters::simulate_parameters()}
-#'   \item \code{parameters::principal_components()}
-#'   \item \code{performance::binned_residuals()}
-#'   \item \code{performance::check_collinearity()}
-#'   \item \code{performance::check_distribution()}
-#'   \item \code{performance::check_heteroscedasticity()}
-#'   \item \code{performance::check_homogeneity()}
-#'   \item \code{performance::check_model()}
-#'   \item \code{performance::check_normality()}
-#'   \item \code{performance::check_outliers()}
-#'   \item \code{performance::compare_performance()}
-#'   \item \code{performance::performance_roc()}
-#' }
 #'
 #' @seealso \href{https://easystats.github.io/see/articles/}{Package-Vignettes}
 #'
@@ -150,15 +118,6 @@ data_plot <- function(x, data = NULL, ...){
 
 
 
-#' @method print data_plot
-#' @importFrom graphics plot
-#' @export
-print.data_plot <- function(x, ...){
-  print(as.data.frame(x))
-}
-
-
-
 #' Complete figure with its attributes
 #'
 #' The \link{data_plot} function usually stores information (such as title, axes labels etc.) as attributes. This function adds those information to the plot.
@@ -187,7 +146,7 @@ print.data_plot <- function(x, ...){
 #' p
 #' p + add_plot_attributes(data)}
 #' @export
-add_plot_attributes <- function(x){
+add_plot_attributes <- function(x) {
   info <- attributes(x)$info
   out <- list(ylab(info$ylab), xlab(info$xlab))